% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R, R/neoantigenR.orig.11082017.R
\name{neoantigenR.get.Model}
\alias{neoantigenR.get.Model}
\alias{neoantigenR.get.Model}
\title{retrieve the predicted gene's exon coordinates and prepare them for visualization}
\usage{
neoantigenR.get.Model()

neoantigenR.get.Model()
}
\value{
a list of gene models will be produced from the GFF file

a list of gene models will be produced from the GFF file
}
\description{
retrieve the predicted gene's exon coordinates and prepare them for visualization

retrieve the predicted gene's exon coordinates and prepare them for visualization
}
