% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R, R/neoantigenR.orig.11082017.R
\name{neoantigenR.initialize}
\alias{neoantigenR.initialize}
\alias{neoantigenR.initialize}
\title{initialize the parameters and functions for neoantigen prediction}
\usage{
neoantigenR.initialize()

neoantigenR.initialize()
}
\value{
global variables will be set after initialization

global variables will be set after initialization
}
\description{
initialize the parameters and functions for neoantigen prediction

initialize the parameters and functions for neoantigen prediction
}
