% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R, R/neoantigenR.orig.11082017.R
\name{neoantigenR.write}
\alias{neoantigenR.write}
\alias{neoantigenR.write}
\title{write the predicted neoantigens for final results}
\usage{
neoantigenR.write()

neoantigenR.write()
}
\value{
a predicted neoantigen file will be written

a predicted neoantigen file will be written
}
\description{
write the predicted neoantigens for final results

write the predicted neoantigens for final results
}
