% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenUtil.R
\name{retrieve.geneName.pacbioID.mapping.v2}
\alias{retrieve.geneName.pacbioID.mapping.v2}
\title{read the gff gene prediction data frame}
\usage{
retrieve.geneName.pacbioID.mapping.v2(gff.file)
}
\arguments{
\item{gff.file}{the name of the pairs intersecting model and reference}
}
\value{
a data frame with annotated namme to gene prediction mapping
}
\description{
read the gff gene prediction data frame
}
